SF: Improve hwrotation handling
Avoid getprop() calls in performance critical display
code paths.
Instead of querying the property each time we need it,
we read it once during initialization and then reuse
the cached value.

This is more appropriate here because we do not expect
the value to change at runtime. In fact, this property
behaves like a compile time constant in the real world:
Set it once and never again (because the angle of your
panel is fixed and does not change after the device
leaves the factory).

Change-Id: I55c4131735a65c7bdde8b00c166913bffa6c4ec3
Signed-off-by: RakeshBatra <raakesh.batra@rediffmail.com>